# frontend/backend-openapi-spec.json
{
	"openapi": "3.0.2",
	"info": { "title": "FastAPI", "version": "0.1.0" },
	"servers": [
		{
			"url": "http://localhost:8000/",
			"description": "Development server"
		}
	],
	"paths": {
...
		"/state": {
			"get": {
				"summary": "Get the current board state.",
				"description": "Returns the current board state as a list of lists of bools.",
				"operationId": "get_board_state",
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {
									"title": "Response Get Board State State Get",
									"type": "array",
									"items": { "type": "array", "items": { "type": "boolean" } }
								}
							}
						}
					}
				}
			},
			"post": {
				"summary": "Set the current board state.",
				"description": "Sets the current board state as a list of lists of bools.",
				"operationId": "set_board_state",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"title": "State",
								"type": "array",
								"items": { "type": "array", "items": { "type": "boolean" } }
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {
									"title": "Response Set Board State State Post",
									"type": "array",
									"items": { "type": "array", "items": { "type": "boolean" } }
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/HTTPValidationError" }
							}
						}
					}
				}
			}
		},
...
	},
	"components": {
		"schemas": {
			"HTTPValidationError": {
				"title": "HTTPValidationError",
				"type": "object",
				"properties": {
					"detail": {
						"title": "Detail",
						"type": "array",
						"items": { "$ref": "#/components/schemas/ValidationError" }
					}
				}
			},
			"ValidationError": {
				"title": "ValidationError",
				"required": ["loc", "msg", "type"],
				"type": "object",
				"properties": {
					"loc": { "title": "Location", "type": "array", "items": { "type": "string" } },
					"msg": { "title": "Message", "type": "string" },
					"type": { "title": "Error Type", "type": "string" }
				}
			}
		}
	}
}